unsplit file
